requirements.txt
absl-py
moviepy
pysrt
pytube3